Implememtatation of Remote AES
This version simplifies AES to only ECB mode and allows running the Cipher1R function in an FPGA as:
Command 2222 - FPGA enters in "read a round key" mode (to be called only once"
Command 1111 - FPGA enters in read input mode (16 bytes) , after this is read it encrypts one round of AES with it and reeturns the ciphertext.

No other commands are implemented.